+++ app/src/main/java/com/example/huascar/shoppingbasket/models/LoyaltyDiscount.java
@@ -8,8 +8,8 @@ import com.example.huascar.shoppingbasket.Interfaces.Discount;
 
 public class LoyaltyDiscount implements Discount {
 
-    public int Discount(int total) {
-        int result = total - (total*2)/100;
+    public double Discount(int total) {
+        double result = total - (total*2)/100;
         return result;
     }
 }
